+++ warehouse/slave_worker_job.go
@@ -135,11 +135,7 @@ func newJobRun(job payload, conf *config.Config, log logger.Logger, stat stats.S
 		encodingFactory: encodingFactory,
 	}
 
-	if conf.IsSet("Warehouse.slaveUploadTimeout") {
-		jr.config.slaveUploadTimeout = conf.GetDuration("Warehouse.slaveUploadTimeout", 10, time.Minute)
-	} else {
-		jr.config.slaveUploadTimeout = conf.GetDuration("Warehouse.slaveUploadTimeoutInMin", 10, time.Minute)
-	}
+	jr.config.slaveUploadTimeout = conf.GetDurationVar(10, time.Minute, "Warehouse.slaveUploadTimeout", "Warehouse.slaveUploadTimeoutInMin")
 	jr.config.numLoadFileUploadWorkers = conf.GetInt("Warehouse.numLoadFileUploadWorkers", 8)
 	jr.config.loadObjectFolder = conf.GetString("WAREHOUSE_BUCKET_LOAD_OBJECTS_FOLDER_NAME", "rudder-warehouse-load-objects")
 
@@ -307,7 +303,7 @@ func (jr *jobRun) uploadLoadFiles(ctx context.Context) ([]uploadResult, error) {
 			jr.config.loadObjectFolder,
 			tableName,
 			jr.job.SourceID,
-			getBucketFolder(jr.job.UniqueLoadGenID, tableName),
+			jr.bucketFolder(jr.job.UniqueLoadGenID, tableName),
 		)
 	}
 
@@ -387,6 +383,10 @@ func (jr *jobRun) uploadLoadFiles(ctx context.Context) ([]uploadResult, error) {
 	}
 
 	return output, nil
+}
+
+func (jr *jobRun) bucketFolder(batchID, tableName string) string {
+	return batchID + "-" + tableName
 }
 
 func (jr *jobRun) reader() (*gzip.Reader, error) {