+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.14.4](https://github.com/rudderlabs/rudder-server/compare/v1.14.3...v1.14.4) (2023-09-29)
+
+
+### Bug Fixes
+
+* use dedup on new records for deltalake ([#3927](https://github.com/rudderlabs/rudder-server/issues/3927)) ([5656ab3](https://github.com/rudderlabs/rudder-server/commit/5656ab3f68dd58e332b68d45f21d1965594ef479))
+
+## [1.14.3](https://github.com/rudderlabs/rudder-server/compare/v1.14.2...v1.14.3) (2023-09-28)
+
+
+### Miscellaneous
+
+* added flags for event audit ([#3924](https://github.com/rudderlabs/rudder-server/issues/3924)) ([6908b0a](https://github.com/rudderlabs/rudder-server/commit/6908b0a03df8c59fd1c22c312a825cb62abdab3d))
+
+## [1.14.2](https://github.com/rudderlabs/rudder-server/compare/v1.14.1...v1.14.2) (2023-09-25)
+
+
+### Bug Fixes
+
+* googlecloudfunction destination implementation and refactor code ([#3907](https://github.com/rudderlabs/rudder-server/issues/3907)) ([0676b0a](https://github.com/rudderlabs/rudder-server/commit/0676b0a70e26d70e1fcb4aeb7fc34c3a258b23ef))
+
 ## [1.14.1](https://github.com/rudderlabs/rudder-server/compare/v1.14.0...v1.14.1) (2023-09-21)